# backend/openapi-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para la gestión de camisetas deportivas, clientes B2B, tallas y ventas. Permite administrar el stock de camisetas, gestionar clientes preferenciales y regulares, calcular precios con descuentos y registrar ventas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
        "/camisetas": {
            "get": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Listar camisetas",
                "description": "Obtiene todas las camisetas registradas junto con sus tallas disponibles.",
                "operationId": "listarCamisetas",
                "responses": {
                    "200": {
                        "description": "Listado de camisetas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Camiseta"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Crear camiseta",
                "description": "Registra una nueva camiseta en el catálogo.",
                "operationId": "crearCamiseta",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CamisetaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Camiseta creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Camiseta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/camisetas/{camiseta}": {
            "get": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Obtener camiseta",
                "description": "Obtiene una camiseta específica y calcula el precio final según el descuento del cliente.",
                "operationId": "obtenerCamiseta",
                "parameters": [
                    {
                        "name": "camiseta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cliente_id",
                        "in": "query",
                        "description": "ID del cliente para aplicar descuento",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Camiseta encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Camiseta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Actualizar camiseta",
                "description": "Actualiza los datos de una camiseta existente.",
                "operationId": "actualizarCamiseta",
                "parameters": [
                    {
                        "name": "camiseta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CamisetaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Camiseta actualizada correctamente"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Eliminar camiseta",
                "description": "Elimina una camiseta siempre que no esté asociada a ventas.",
                "operationId": "eliminarCamiseta",
                "parameters": [
                    {
                        "name": "camiseta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Camiseta eliminada correctamente"
                    },
                    "409": {
                        "description": "La camiseta posee ventas asociadas"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Camiseta": {
                "title": "Camiseta",
                "description": "Modelo completo de camiseta deportiva",
                "required": [
                    "titulo",
                    "club",
                    "pais",
                    "tipo",
                    "color",
                    "precio",
                    "stock",
                    "codigo_producto"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Camiseta Local 2025 - Selección Chilena"
                    },
                    "club": {
                        "type": "string",
                        "example": "Selección Chilena"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Chile"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Local"
                    },
                    "color": {
                        "type": "string",
                        "example": "Rojo"
                    },
                    "precio": {
                        "type": "integer",
                        "example": 45000
                    },
                    "precio_final": {
                        "description": "Precio calculado según el descuento aplicado al cliente consultante",
                        "type": "integer",
                        "example": 40500,
                        "nullable": true
                    },
                    "stock": {
                        "type": "integer",
                        "example": 20
                    },
                    "detalles": {
                        "type": "string",
                        "example": "Edición aniversario 2025",
                        "nullable": true
                    },
                    "codigo_producto": {
                        "type": "string",
                        "example": "SCL2025L"
                    },
                    "tallas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Talla"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CamisetaInput": {
                "title": "Camiseta Input",
                "description": "Datos para crear o actualizar una camiseta",
                "required": [
                    "titulo",
                    "club",
                    "pais",
                    "tipo",
                    "color",
                    "precio",
                    "stock",
                    "codigo_producto"
                ],
                "properties": {
                    "titulo": {
                        "type": "string",
                        "example": "Camiseta Local 2025 - Selección Chilena"
                    },
                    "club": {
                        "type": "string",
                        "example": "Selección Chilena"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Chile"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Local"
                    },
                    "color": {
                        "type": "string",
                        "example": "Rojo"
                    },
                    "precio": {
                        "type": "integer",
                        "example": 45000
                    },
                    "stock": {
                        "type": "integer",
                        "example": 20
                    },
                    "detalles": {
                        "type": "string",
                        "example": "Edición aniversario 2025",
                        "nullable": true
                    },
                    "codigo_producto": {
                        "type": "string",
                        "example": "SCL2025L"
                    },
                    "tallas": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Talla": {
                "title": "Talla",
                "description": "Modelo completo de talla disponible para camisetas",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "M"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}